:sparkles: #107 내 픽 올리기, 쿡찌르기 유스케이스 MainViewModel로 이동
