+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/ui/main/MainActivity.kt
@@ -16,15 +16,18 @@ import com.mashup.gabbangzip.sharedalbum.presentation.ui.eventcreation.EventCrea
 import com.mashup.gabbangzip.sharedalbum.presentation.ui.login.LoginActivity
 import com.mashup.gabbangzip.sharedalbum.presentation.ui.main.navigation.MainNavHost
 import com.mashup.gabbangzip.sharedalbum.presentation.ui.main.navigation.MainRoute
+import com.mashup.gabbangzip.sharedalbum.presentation.utils.PicPhotoPicker
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     private val viewModel by viewModels<MainViewModel>()
+    private lateinit var photoPicker: PicPhotoPicker
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         viewModel.registerFcmToken()
+        initPhotoPicker()
 
         setContent {
             SharedAlbumTheme {
@@ -50,8 +53,19 @@ class MainActivity : ComponentActivity() {
                             LoginActivity.openActivity(this@MainActivity)
                             finish()
                         },
+                        onClickOpenPhotoPickerButton = {
+                            photoPicker.open()
+                        },
                     )
                 }
+            }
+        }
+    }
+
+    private fun initPhotoPicker() {
+        photoPicker = PicPhotoPicker.create(this) { uri ->
+            if (uri != null) {
+                viewModel.uploadMyPic(uri)
             }
         }
     }
